requirements.txt
alembic==1.14.0
APScheduler==3.10.4
attrs==24.2.0
blinker==1.9.0
build==1.2.2.post1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
et_xmlfile==2.0.0
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.7
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip-tools==7.4.1
pyotp==2.9.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
selenium==4.26.1
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
zope.event==5.0
zope.interface==7.1.1
